refactor(navigation): partir VinilosNavHost en grafos por feature (B2)
VinilosNavHost (~318 -> 56 lineas) partido en albumGraph/artistGraph/
collectorGraph/prizeGraph (extensiones NavGraphBuilder) + selectedTab
puro testeable en JVM. Movimiento verbatim de cada bloque composable,
sin cambio de rutas/args/keys/logica.
